app/api/requirements.txt
langchain
langchain_experimental
langgraph
langchain-openai
python-dotenv  # load_dotenv comes from python-dotenv package
tiktoken
sentence-transformers
pymupdf
langchain_pymupdf4llm
langchain-community
playwright
lxml
gradio
rank_bm25
ollama==0.5.3
selenium
requests==2.32.5
docling
easyocr==1.7.2
pdf2image==1.17.0
llama-index-embeddings-huggingface==0.6.0
llama-index-postprocessor-flag-embedding-reranker==0.4.0
chromadb==1.0.16
llama-index==0.13.1
git+https://github.com/FlagOpen/FlagEmbedding.git@1.3.5
llama-index-vector-stores-chroma==0.5.0
llama-index-retrievers-bm25==0.6.3
